fixed id increment and homepage

+++ finalProject/untitled/src/main/java/quizpackage/Account.java
@@ -6,5 +6,5 @@ public interface Account {
     String getUsername();
     String getPassword();
     int getAge();
-    long getId();
+    int getId();
 }
